+++ DLStarRating/UIView+Subviews.h
@@ -15,6 +15,6 @@
 
 @interface UIView (Subviews)
 
-- (UIView*)subViewWithTag:(int)tag;
+- (UIView*)subViewWithTag:(NSInteger)tag;
 
 @end
