+++ DLStarRating/UIView+Subviews.m
@@ -15,7 +15,7 @@
 
 @implementation UIView (Subviews)
 
-- (UIView*)subViewWithTag:(int)tag {
+- (UIView*)subViewWithTag:(NSInteger)tag {
 	for (UIView *v in self.subviews) {
 		if (v.tag == tag) {
 			return v;
